+++ src/main/kotlin/dev/mizarc/waystonewarps/interaction/messaging/ColorPalettes.kt
@@ -3,12 +3,14 @@ package dev.mizarc.waystonewarps.interaction.messaging
 import net.kyori.adventure.text.format.TextColor
 
 enum class PrimaryColourPalette(val color: TextColor?) {
+    PRIMARY(TextColor.fromHexString("#ffaa00")),
     INFO(TextColor.fromHexString("#589bbf")),
     SUCCESS(TextColor.fromHexString("#58bf78")),
     PENDING(TextColor.fromHexString("#e0d01e")),
     CANCELLED(TextColor.fromHexString("#d3270d")),
     FAILED(TextColor.fromHexString("#910f0f")),
-    SPECIAL(TextColor.fromHexString("#980dd3"))
+    SPECIAL(TextColor.fromHexString("#980dd3")),
+    UNAVAILABLE(TextColor.fromHexString("#808080"))
 }
 
 enum class AccentColourPalette(val color: TextColor?) {